RealTime-DeepfakeDetection-in-the-RealWorld/requirements.txt
albumentations==1.4.0
appdirs==1.4.4
bagnets @ git+https://github.com/wielandbrendel/bag-of-local-features-models.git@8493fb442c9b9ccfeaa970c950ca67832017f072
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
faiss-gpu==1.7.2
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.2
pillow==10.2.0
protobuf==4.25.2
psutil==5.9.8
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==1.40.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.1.30
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
wandb==0.16.3
xgboost==2.0.3
zipp==3.17.0
